🚚 rename: performance query 위계 변경
